pyproject.toml
[tool.poetry]
name = "geekbrains-home-flask"
version = "0.0.0"
description = ""
authors = ["vios <i.svoboda.vios@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "geekbrains_home_flask"}]

[tool.poetry.dependencies]
python = "3.10.8"
Flask = "2.2.2"
flask-sqlalchemy = "3.0.2"
flask-login = "0.6.2"
flask-migrate = "4.0.0"
psycopg2 = "2.9.5"
bcrypt-flask = "1.0.2"
wtform = "1.0"
email-validator = "1.3.0"
flask-wtf = "1.1.0"
flask-admin = "1.6.0"
flask-restful = "0.3.9"
pyyaml = "5.3.1"
flask-apispec = "0.11.4"
flask-marshmallow = "0.14.0"
marshmallow-sqlalchemy = "0.28.1"
flask-perm = "0.2.8"

[tool.poetry.group.dev.dependencies]
pyvim = "3.0.3"
pynvim = "0.4.3"
black = "22.12.0"
neovim = "0.3.1"
aiohttp = "3.8.3"
jedi = "0.18.2"
ipdb = "0.13.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
